+++ R/buttons.R
@@ -16,7 +16,6 @@
 #'   \item{\code{relative_size}}
 #'   \item{\code{add_nodes}}
 #'   \item{\code{add_edges}}
-#'   \item{\code{add_nodes_edges}}
 #'   \item{\code{drop_nodes}}
 #'   \item{\code{drop_edges}}
 #'   \item{\code{animate}}
@@ -24,15 +23,19 @@
 #'   \item{\code{export_img}}
 #' }
 #' 
+#' @details You can pass multiple events as a vector, see examples.
+#' 
 #' @examples 
 #' nodes <- sg_make_nodes() 
-#' edges <- sg_make_edges(nodes, 17)
+#' edges <- sg_make_edges(nodes)
 #' 
+#' # Button starts the layout and stops it after 3 seconds
 #' sigmajs() %>% 
 #'   sg_nodes(nodes, id, size) %>% 
 #'   sg_edges(edges, id, source, target) %>% 
 #'   sg_force_start() %>% 
-#'   sg_button("start layout", "force_start")
+#'   sg_force_stop(3000) %>% 
+#'   sg_button("start layout", c("force_start", "force_stop"))
 #'   
 #' # additional nodes
 #' nodes2 <- sg_make_nodes()
@@ -46,7 +49,7 @@
 #'   sg_add_nodes(nodes2, delay, id, label, size, color) %>% 
 #'   sg_button("add nodes", "add_nodes")
 #' 
-#' @note The default class (\code{btn btn-default}) works with Bootstrap 3, the default framework for Shiny and R markdown.
+#' @note The default class (\code{btn btn-default}) works with Bootstrap 3 (the default framework for Shiny and R markdown).
 #' 
 #' @export
 sg_button <- function(sg, label, event, class = "btn btn-default"){
@@ -56,8 +59,11 @@ sg_button <- function(sg, label, event, class = "btn btn-default"){
   
   .test_sg(sg)
   
-  if(!event %in% .valid_events())
-    stop("incorrect event", call. = FALSE)
+  for(ev in event)
+    if(!ev %in% .valid_events()) stop(paste(ev, "is not a known event"), call. = FALSE)
+  
+  if("add_nodes_edges" %in% event)
+    warning("add_nodes_edges is deprecated, in favour of c('add_nodes', 'add_edges')", call. = FALSE)
   
   btn <- list(
     label = label,